Bug fixes and improvments
- Fixed a bug where the class name of the main JavaScript class was not
properly camelcased, causing compilation errors.
- Improve the singletonian nature of the main js class by adding a
getInstance method. The getInstance method returns the unique instance
of the class if it exists, or create it if it doesn’t and return it. Any
direct call to the main class constructor will return the unique
instance.
- Main class init method now can’t be called more than one time.
- Now if changes are made to Gruntfile.js while the watch taks is
running, the configuration automatically reload to take your changes
into account.